+++ CHANGELOG.md
@@ -15,6 +15,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix: Update `k-inputSet` vertical spacing.
 - Fix: Update `k-RadioButtonSet` line-height.
 
+## [23.17.0] - 2018-12-06
+
+Features:
+- Add `links` prop on `TeamCard` component.
+  `socialLink` is deprecated.
+- Add `k-u-hidden@{viewportRange}` utilities.
+
+Fix:
+- `Marger` use `StyleRoot` only if `top` or `bottom` prop are an object.
+
 ## [23.16.1] - 2018-11-30
 
 Fix: